Add doc generation for users that ended course

+++ src/main/java/com/bsuir/trainingcenter/service/CertificateService.java
@@ -6,7 +6,7 @@ import com.bsuir.trainingcenter.entity.view.CertificateInfoView;
 import java.util.List;
 
 public interface CertificateService {
-    CertificateInfoView getCertificateInfo(int id);
+    CertificateInfoView getCertificateInfo(long id);
 
     boolean addCertificate(Certificate certificate);
